third_party/WebKit/LayoutTests/fast/css/style-scoped/style-scoped-scoping-nodes-different-order-expected.txt

Test a case whether order of scoping nodes is different from order of style elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.styleSheets.length is 4
PASS document.styleSheets[1].cssRules[0].cssText is "span { color: green; }"
PASS document.styleSheets[2].cssRules[0].cssText is "span { color: grey; }"
PASS document.styleSheets[3].cssRules[0].cssText is "span { border: 1px solid green; }"
FAIL borderColor("target") should be rgb(0, 128, 0). Was rgb(128, 128, 128).
PASS successfullyParsed is true

TEST COMPLETE
